*项目建设 1.实现了一个GreenDao访问接口和实现类

+++ app/src/main/java/com/example/flink/item/StickyNoteItem.java
@@ -88,7 +88,7 @@ public class StickyNoteItem {
             return this;
         }
 
-        public Builder setOrder(int order) {
+        public Builder setOrder(long order) {
             this.order = order;
             return this;
         }
